+++ ViviEstu_api/src/main/java/com/viviestu/viviestu_api/dto/CalificacionRequest.java
@@ -1,13 +1,13 @@
 package com.viviestu.viviestu_api.dto;
 
 public class CalificacionRequest {
-    private Integer usuarioId;
+    private Long usuarioId;
     private Integer puntuacion; // 1-5
 
     public CalificacionRequest() {}
 
-    public Integer getUsuarioId() { return usuarioId; }
-    public void setUsuarioId(Integer usuarioId) { this.usuarioId = usuarioId; }
+    public Long getUsuarioId() { return usuarioId; }
+    public void setUsuarioId(Long usuarioId) { this.usuarioId = usuarioId; }
 
     public Integer getPuntuacion() { return puntuacion; }
     public void setPuntuacion(Integer puntuacion) { this.puntuacion = puntuacion; }
